代码重写，增加功能
