+++ run.conf
@@ -1,0 +1,18 @@
+[mode]
+run_mode = flow
+read_all = True
+pcap_loc = ./
+pcap_name = test.pcap
+csv_name = feature.csv
+
+
+[feature]
+print_port = False
+print_colname = True
+add_tag = False
+
+
+[joblib]
+dump_switch = False
+load_switch = False
+load_name = flows.data
